+++ packages/file-upload-item/src/context/file-upload-item-context.ts
@@ -10,6 +10,7 @@ interface ContextType
         | 'error'
         | 'title'
         | 'subtitle'
+        | 'subtitleColor'
         | 'uploadDate'
         | 'size'
         | 'id'