+++ packages/action-button/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @alfalab/core-components-action-button
 
+## 3.0.7
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.3
+- @alfalab/core-components-spinner@6.0.7
+
+## 3.0.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.2
+- @alfalab/core-components-spinner@6.0.6
+
 ## 3.0.5
 
 ### Patch Changes