+++ packages/chart/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-chart
 
+## 5.0.10
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.10
+
+## 5.0.9
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.9
+
 ## 5.0.8
 
 ### Patch Changes
@@ -603,8 +619,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Изменена директория расположения индексных файлов в опубликованных пакетах (удалена
   директория dist)
-
-Co-authored-by: Vladimir Gevak <VGevak@alfabank.ru>
 
 ## [2.2.1](https://github.com/core-ds/core-components/compare/@alfalab/core-components-chart@2.2.0...@alfalab/core-components-chart@2.2.1) (2022-08-17)