+++ packages/form-control/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @alfalab/core-components-form-control
 
+## 14.0.9
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.3
+- @alfalab/core-components-mq@6.0.7
+
+## 14.0.8
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.2
+- @alfalab/core-components-mq@6.0.6
+
 ## 14.0.7
 
 ### Patch Changes
@@ -709,8 +727,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Изменена директория расположения индексных файлов в опубликованных пакетах (удалена
   директория dist)
-
-Co-authored-by: Vladimir Gevak <VGevak@alfabank.ru>
 
 ## [7.6.1](https://github.com/core-ds/core-components/compare/@alfalab/core-components-form-control@7.6.0...@alfalab/core-components-form-control@7.6.1) (2022-08-17)