+++ packages/icon-button/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @alfalab/core-components-icon-button
 
+## 8.0.10
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-button@13.2.2
+- @alfalab/core-components-mq@6.0.7
+
+## 8.0.9
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-button@13.2.1
+- @alfalab/core-components-mq@6.0.6
+
 ## 8.0.8
 
 ### Patch Changes
@@ -817,8 +835,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Изменена директория расположения индексных файлов в опубликованных пакетах (удалена
   директория dist)
-
-Co-authored-by: Vladimir Gevak <VGevak@alfabank.ru>
 
 ## [4.3.2](https://github.com/core-ds/core-components/compare/@alfalab/core-components-icon-button@4.3.1...@alfalab/core-components-icon-button@4.3.2) (2022-08-17)