+++ packages/mq/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-mq
 
+## 6.0.7
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.3
+
+## 6.0.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.2
+
 ## 6.0.5
 
 ### Patch Changes
@@ -233,8 +249,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Изменена директория расположения индексных файлов в опубликованных пакетах (удалена
   директория dist)
-
-Co-authored-by: Vladimir Gevak <VGevak@alfabank.ru>
 
 ## [3.1.1](https://github.com/core-ds/core-components/compare/@alfalab/core-components-mq@3.1.0...@alfalab/core-components-mq@3.1.1) (2022-08-17)