+++ packages/toast-plate/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @alfalab/core-components-toast-plate
 
+## 9.1.6
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.3
+- @alfalab/core-components-mq@6.0.7
+- @alfalab/core-components-icon-button@8.0.10
+
+## 9.1.5
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-shared@2.2.2
+- @alfalab/core-components-mq@6.0.6
+- @alfalab/core-components-icon-button@8.0.9
+
 ## 9.1.4
 
 ### Patch Changes
@@ -1087,8 +1107,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Изменена директория расположения индексных файлов в опубликованных пакетах (удалена
   директория dist)
-
-Co-authored-by: Vladimir Gevak <VGevak@alfabank.ru>
 
 ## [4.4.2](https://github.com/core-ds/core-components/compare/@alfalab/core-components-toast-plate@4.4.1...@alfalab/core-components-toast-plate@4.4.2) (2022-08-17)